Refactored the way the HttpRequestInitializer is created for Authed- and non-authed Api calls is created
Added an ExceptionHandler system to ApiCalls. Any call can now listen to Exceptions by calling .onException() and passing a handler before executing a call.

The activity now checks whether the server is reachable and displays a Toast if not

Added a new dependency to Toaster, which allows us to send Toasts from different threads (yey)